Showing Incorrect Output for Checking MAC change
The program displayed '[-] MAC Address did not get changed' even though the MAC Address got changed. The error was that the new MAC Address was not updated in the variable 'current_mac' after attempting to change the MAC Address using the method 'change_mac()'. Hence after attempting to change the MAC Address, the value of 'current_mac' is updated.